+++ src/Pumukit/SchemaBundle/Document/MediaType/Metadata/VideoAudio.php
@@ -163,4 +163,9 @@ final class VideoAudio implements MediaMetadata
     {
         return json_decode($this->metadata(), false, 512, JSON_THROW_ON_ERROR);
     }
+
+    public function isEmpty(): bool
+    {
+        return empty($this->metadata);
+    }
 }